requirements.txt
pyyaml
requests
structlog
psutil
